server/mqtt_image/requirements.txt
astroid==2.3.3
Events==0.3
googleapis-common-protos==1.51.0
grpcio==1.27.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
paho-mqtt==1.5.0
protobuf==3.11.3
pylint==2.4.4
rope==0.16.0
six==1.14.0
ttn==2.1.5
typed-ast==1.4.1
wrapt==1.11.2
